Added download list and bash profile files
This covers up to chapter 4 section 4 of the LFS book. Plus, when doing
this on another machine, it is necessary to backtrack and mount the
drive as /mnt/lfs, export it to the $LFS variabel, add an soft link
pointing /tools to the root directory, setting up the lfs user and
group, and copying the bash_profile and bashrc files to its home
directory.